refactor: 提取 configService 从 workspaceModuleService (Phase 3)
创建独立的 configService.ts (~309行)，包含所有纯 YAML 配置读写函数：
serializeConfig, parseYamlConfig, parseLegacyConfig, initializeConfig,
loadConfig, writeConfig, withAppliedModule, withoutModule 等 13 个函数。

workspaceModuleService.ts 从 ~1394 行减至 ~1203 行，配置相关方法改为
单行委托到 configService。公共 API 完全不变，所有 352 测试通过。
